no errors left, added ci. and phpdoc.yml
