+++ phpunit.xml.dist
@@ -1,11 +1,12 @@
-<phpunit bootstrap="test/Horde/Share/bootstrap.php" colors="true">
-  <filter>
-    <whitelist>
+<?xml version="1.0"?>
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" bootstrap="test/Horde/Share/bootstrap.php" colors="true" xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.3/phpunit.xsd">
+  <coverage>
+    <include>
       <directory suffix=".php">lib</directory>
-    </whitelist>
-  </filter>
+    </include>
+  </coverage>
   <testsuites>
-    <testsuite>
+    <testsuite name="horde/Share">
       <directory>test</directory>
     </testsuite>
   </testsuites>
