+++ test/Horde/Share/Sql/MysqliTest.php
@@ -2,11 +2,7 @@
 /**
  * Prepare the test setup.
  */
-namespace Horde\Share;
-use \Sql;
-use Horde_Share_Test_Sql_Base as Base;
-
-require_once __DIR__ . '/Base.php';
+namespace Horde\Share\Sql;
 
 /**
  * Copyright 2010-2017 Horde LLC (http://www.horde.org/)
@@ -17,7 +13,7 @@ require_once __DIR__ . '/Base.php';
  * @subpackage UnitTests
  * @license    http://www.horde.org/licenses/lgpl21 LGPL 2.1
  */
-class MysqliTest extends Base
+class MysqliTest extends BaseTestCase
 {
     public static function setUpBeforeClass(): void
     {
