+++ test/Horde/Share/Sql/Pdo/SqliteTest.php
@@ -2,13 +2,15 @@
 /**
  * Prepare the test setup.
  */
-namespace Horde\Share;
-use \Sql;
-use \Pdo;
-use Horde_Share_Test_Sql_Base as Base;
+namespace Horde\Share\Sql\Pdo;
+use Horde\Share\Sql\BaseTestCase;
 use \Horde_Test_Factory_Db;
-
-require_once __DIR__ . '/../Base.php';
+use \Horde_Share_Stub_Group;
+use \PDO;
+use Horde\Share\Horde_Perms;
+use \Horde_Support_Stub;
+use \Horde_Db_Migration_Base;
+use \Horde_Injector;
 
 /**
  * Copyright 2010-2017 Horde LLC (http://www.horde.org/)
@@ -19,17 +21,17 @@ require_once __DIR__ . '/../Base.php';
  * @subpackage UnitTests
  * @license    http://www.horde.org/licenses/lgpl21 LGPL 2.1
  */
-class SqliteTest extends Base
+class SqliteTest extends BaseTestCase
 {
     public static function setUpBeforeClass(): void
     {
         $factory_db = new Horde_Test_Factory_Db();
 
-        try {
+        if (class_exists(Horde_Injector::class)) {
             self::$db = $factory_db->create();
             parent::setUpBeforeClass();
-        } catch (Horde_Test_Exception $e) {
-            self::$reason = 'Sqlite not available';
+        } else {
+            self::$reason = 'Horde_Injector not available';
         }
     }
 }
