+++ test/Horde/Share/Sqlng/Pdo/SqliteTest.php
@@ -3,10 +3,8 @@
  * Prepare the test setup.
  */
 namespace Horde\Share\Sqlng\Pdo;
-use Horde_Share_Test_Sqlng_Base as Base;
+use Horde\Share\Sqlng\BaseTestCase;
 use \Horde_Test_Factory_Db;
-
-require_once __DIR__ . '/../Base.php';
 
 /**
  * Copyright 2010-2017 Horde LLC (http://www.horde.org/)
@@ -17,17 +15,17 @@ require_once __DIR__ . '/../Base.php';
  * @subpackage UnitTests
  * @license    http://www.horde.org/licenses/lgpl21 LGPL 2.1
  */
-class SqliteTest extends Base
+class SqliteTest extends BaseTestCase
 {
     public static function setUpBeforeClass(): void
     {
         $factory_db = new Horde_Test_Factory_Db();
 
-        try {
+        if (class_exists(Horde_Injector::class)) {
             self::$db = $factory_db->create();
             parent::setUpBeforeClass();
-        } catch (Horde_Test_Exception $e) {
-            self::$reason = 'Sqlite not available';
+        } else {
+            self::$reason = 'Horde_Injector not available';
         }
     }
 }
