+++ test/Horde/Share/bootstrap.php
@@ -1,3 +1,5 @@
 <?php
-require_once 'Horde/Test/Bootstrap.php';
+if (!class_exists('Horde_Test_Bootstrap')) {
+    		require_once 'Horde/Test/Bootstrap.php';
+}
 Horde_Test_Bootstrap::bootstrap(dirname(__FILE__));
